+++ demo/whitelist.js
@@ -107,9 +107,9 @@ function readFile() {
       let sellValid = isValidDate(data[1])
       let buyValid = isValidDate(data[2])
       let kycExpiryDate = isValidDate(data[3])
+      let canBuyFromSTO = (typeof JSON.parse(data[4].toLowerCase())) == "boolean" ? JSON.parse(data[4].toLowerCase()) : "not-valid";
 
-
-      if (isAddress && sellValid && buyValid && kycExpiryDate) {
+      if (isAddress && sellValid && buyValid && kycExpiryDate && (canBuyFromSTO != "not-valid") ) {
         let userArray = new Array()
         let checksummedAddress = web3.utils.toChecksumAddress(data[0]);
 
@@ -117,6 +117,7 @@ function readFile() {
         userArray.push(sellValid)
         userArray.push(buyValid)
         userArray.push(kycExpiryDate)
+        userArray.push(canBuyFromSTO)
         // console.log(userArray)
         allocData.push(userArray);
         fullFileData.push(userArray);
@@ -137,6 +138,7 @@ function readFile() {
         userArray.push(sellValid)
         userArray.push(buyValid)
         userArray.push(kycExpiryDate);
+        userArray.push(canBuyFromSTO);
         badData.push(userArray);
         fullFileData.push(userArray)
       }
@@ -188,6 +190,7 @@ async function setInvestors() {
       let fromTimesArray = [];
       let toTimesArray = [];
       let expiryTimeArray = [];
+      let canBuyFromSTOArray = [];
 
       //splitting the user arrays to be organized by input
       for (let j = 0; j < distribData[i].length; j++) {
@@ -195,12 +198,13 @@ async function setInvestors() {
         fromTimesArray.push(distribData[i][j][1])
         toTimesArray.push(distribData[i][j][2])
         expiryTimeArray.push(distribData[i][j][3])
+        canBuyFromSTOArray.push(distribData[i][j][4])
       }
 
       //fromTimes is ability to sell coin FROM your account (2nd row in csv, 2nd parameter in modifyWhiteList() )
       //toTimes is ability to buy coins TOwards your account (3rd row in csv, 3rd parameter in modifyWhiteList() )
       //expiryTime is time at which KYC of investor get expired (4th row in csv, 4rd parameter in modifyWhiteList() )
-      let r = await generalTransferManager.methods.modifyWhitelistMulti(investorArray, fromTimesArray, toTimesArray, expiryTimeArray).send({ from: Issuer, gas: 5000000, gasPrice: DEFAULT_GAS_PRICE })
+      let r = await generalTransferManager.methods.modifyWhitelistMulti(investorArray, fromTimesArray, toTimesArray, expiryTimeArray, canBuyFromSTOArray).send({ from: Issuer, gas: 5000000, gasPrice: DEFAULT_GAS_PRICE })
       console.log(`Batch ${i} - Attempting to modifyWhitelist accounts:\n\n`, investorArray, "\n\n");
       console.log("---------- ---------- ---------- ---------- ---------- ---------- ---------- ----------");
       console.log("Whitelist transaxction was successful.", r.gasUsed, "gas used. Spent:", web3.utils.fromWei(BigNumber(r.gasUsed * DEFAULT_GAS_PRICE).toString(), "ether"), "Ether");
@@ -235,12 +239,14 @@ async function setInvestors() {
     let fromTime_Event = event_data[i].returnValues._fromTime
     let toTime_Event = event_data[i].returnValues._toTime
     let expiryTime_Event = event_data[i].returnValues._expiryTime
+    let canBuyFromSTO_Event = event_data[i].returnValues._canBuyFromSTO
     let blockNumber = event_data[i].blockNumber
 
     combineArray.push(investorAddress_Event);
     combineArray.push(fromTime_Event);
     combineArray.push(toTime_Event);
     combineArray.push(expiryTime_Event);
+    combineArray.push(canBuyFromSTO_Event);
     combineArray.push(blockNumber)
 
     investorData_Events.push(combineArray)
@@ -250,7 +256,7 @@ async function setInvestors() {
 
       //the block number form the event we are checking is bigger, so we gotta replace it
       if (investorObjectLookup[investorAddress_Event].recordedBlockNumber < blockNumber) {
-        investorObjectLookup[investorAddress_Event] = { fromTime: fromTime_Event, toTime: toTime_Event, expiryTime: expiryTime_Event, recordedBlockNumber: blockNumber };
+        investorObjectLookup[investorAddress_Event] = { fromTime: fromTime_Event, toTime: toTime_Event, expiryTime: expiryTime_Event, canBuyFromSTO: canBuyFromSTO_Event, recordedBlockNumber: blockNumber };
         updatedInvestors += 1;
         // investorAddress_Events.push(investorAddress_Event); not needed, because we start the obj with zero events
 
@@ -259,7 +265,7 @@ async function setInvestors() {
       }
       //we have never recorded this address as an object key, so we need to add it to our list of investors updated by the csv
     } else {
-      investorObjectLookup[investorAddress_Event] = { fromTime: fromTime_Event, toTime: toTime_Event, expiryTime: expiryTime_Event, recordedBlockNumber: blockNumber };
+      investorObjectLookup[investorAddress_Event] = { fromTime: fromTime_Event, toTime: toTime_Event, expiryTime: expiryTime_Event, canBuyFromSTO: canBuyFromSTO_Event, recordedBlockNumber: blockNumber };
       totalInvestors += 1;
       // investorAddress_Events.push(investorAddress_Event);
     }