+++ test/security_token.js
@@ -292,6 +292,7 @@ contract('SecurityToken', accounts => {
                 fromTime,
                 toTime,
                 expiryTime,
+                true,
                 {
                     from: token_owner,
                     gas: 6000000
@@ -324,6 +325,7 @@ contract('SecurityToken', accounts => {
                 fromTime,
                 toTime,
                 expiryTime,
+                true,
                 {
                     from: token_owner,
                     gas: 6000000
@@ -505,6 +507,7 @@ contract('SecurityToken', accounts => {
                     fromTime,
                     toTime,
                     expiryTime,
+                    true,
                     {
                         from: token_owner,
                         gas: 6000000
@@ -652,6 +655,7 @@ contract('SecurityToken', accounts => {
                     fromTime,
                     toTime,
                     expiryTime,
+                    true,
                     {
                         from: account_issuer,
                         gas: 500000
@@ -680,6 +684,7 @@ contract('SecurityToken', accounts => {
                     fromTime,
                     toTime,
                     expiryTime,
+                    true,
                     {
                         from: account_issuer,
                         gas: 500000
@@ -716,6 +721,7 @@ contract('SecurityToken', accounts => {
                     fromTime,
                     toTime,
                     expiryTime,
+                    true,
                     {
                         from: account_delegate,
                         gas: 6000000
@@ -756,6 +762,7 @@ contract('SecurityToken', accounts => {
                     0,
                     0,
                     0,
+                    true,
                     {
                         from: account_delegate,
                         gas: 6000000
@@ -794,6 +801,7 @@ contract('SecurityToken', accounts => {
                 fromTime,
                 toTime,
                 expiryTime,
+                true,
                 {
                     from: account_delegate,
                     gas: 6000000