+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file.
 * Module Factories now allow their owner to set 3 types of fees: Setup fee, Usage fee, Monthly fee.
 * Added Checkpoint feature to token. This allows the issuer to create snapshots of the token's balances and totalSupply to be used for casting votes or distributing dividends. Created CLI to demo this feature.
 * Forever minting is prevented by adding one more check for `mintingFinished` flag. Issuer could stop minting by calling the `finishMinting()` function. 
+* Issuer can prevent addresses from participating in the STO. To facilitate this, `_canBuyFromSTO` has been added to modifyWhitelist function.
+* Added multi-mint feature.
+
+
 
 ## Changed